feat: Add currentSession query to retrieve the current user session
This commit adds a new query called currentSession to the user/query.ts file. The query retrieves the current user session by using the getAuthSessionId function from the @convex-dev/auth/server package. If the session ID is null, the query returns null. Otherwise, it fetches the session data from the database and returns it. This new query enhances the functionality of the user module by providing access to the current user session information.